requirements.txt
pandas
numpy
pyyaml
scikit-learn
lightgbm; sys_platform != 'darwin' or platform_machine != 'arm64'
catboost
fastapi
uvicorn
joblib
pyarrow
